+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ### Added
 - Closes #68, adds ability to select base price from buy, sell or split prices.
 - Added help icons and explanations for item sheet
+- Closes #71, Ability to select Immediate pricing over top 5% average
 
 ### Changed
+- Updated some readme text
 
 ### Fixed
 - Fixed mistake in displaying total raw values for refined ore in calculation sheets.